pyproject.toml
[project]
name = "project"
version = "0.0.1"
description = "Project"
authors = [{ name = "Dmitry", email = "dbevzpr@mail.ru" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.11"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.11",
    "fastapi-sqlalchemy-toolkit>=0.7.12",
    "networkx>=3.4.2",
    "pydantic-settings>=2.8.1",
    "redis>=5.2.1",
    "sqlmodel>=0.0.24",
    "uvicorn[standard]>=0.34.0",
]


[dependency-groups]
dev = []




[tool.pytest.ini_options]
testpaths = ["tests"]
# asyncio_mode = "auto"

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true
